+++ ResumableFunctions.Handler/Core/Abstraction/IFirstWaitProcessor.cs
@@ -5,6 +5,6 @@ namespace ResumableFunctions.Handler.Core.Abstraction
     public interface IFirstWaitProcessor
     {
         Task<MethodWait> CloneFirstWait(MethodWait firstMatchedMethodWait);
-        Task RegisterFirstWait(long functionId);
+        Task RegisterFirstWait(int functionId);
     }
 }
