+++ ResumableFunctions.Handler/DataAccess/Abstraction/IPushedCallsRepo.cs
@@ -4,6 +4,6 @@ namespace ResumableFunctions.Handler.DataAccess.Abstraction;
 
 public interface IPushedCallsRepo
 {
-    Task<PushedCall> GetById(long pushedCallId);
+    Task<PushedCall> GetById(int pushedCallId);
     void Add(PushedCall pushedCall);
 }
