+++ ResumableFunctions.Handler/InOuts/WaitMethodIdentifier.cs
@@ -5,7 +5,7 @@ public  class WaitMethodIdentifier : MethodIdentifier
     public bool CanPublishFromExternal { get; internal set; }
 
     public MethodsGroup MethodGroup { get; internal set; }
-    public long MethodGroupId { get; internal set; }
+    public int MethodGroupId { get; internal set; }
 
     internal override void FillFromMethodData(MethodData methodData)
     {
