+++ ResumableFunctions.Handler/UiService/InOuts/FunctionInstanceDetails.cs
@@ -7,7 +7,7 @@ namespace ResumableFunctions.Handler.UiService.InOuts
     {
         public string FunctionUrn { get;}
         public string FunctionName { get;}
-        public long InstanceId { get; }
+        public int InstanceId { get; }
         public FunctionStatus Status { get; }
         public string InstanceData { get; }
         public DateTime Created { get; }
@@ -16,10 +16,10 @@ namespace ResumableFunctions.Handler.UiService.InOuts
         public ArrayList Waits { get; }
         public List<MethodWaitDetails> MethodWaitDetails { get; }
         public List<LogRecord> Logs { get; }
-        public long FunctionId { get; }
+        public int FunctionId { get; }
 
         public FunctionInstanceDetails(
-            long instanceId,long functionId, string name, string functionName, FunctionStatus status, string instanceData, DateTime created, DateTime modified, int errorsCount, ArrayList waits, List<LogRecord> logs)
+            int instanceId,int functionId, string name, string functionName, FunctionStatus status, string instanceData, DateTime created, DateTime modified, int errorsCount, ArrayList waits, List<LogRecord> logs)
         {
             InstanceId = instanceId;
             FunctionId = functionId;
