+++ ResumableFunctions.Handler/UiService/InOuts/ServiceInfo.cs
@@ -1,6 +1,6 @@
 ﻿namespace ResumableFunctions.Handler.UiService.InOuts
 {
-    public record ServiceInfo(long Id, string Name, string Url,string[] Dlls,DateTime Registration,DateTime LastScan)
+    public record ServiceInfo(int Id, string Name, string Url,string[] Dlls,DateTime Registration,DateTime LastScan)
     {
         public int LogErrors { get; set; }
         public int FunctionsCount { get; set; }
